prevent cookie conflicts when running the server and client on the same domain, by removing the session from delivery routes. (#52)
