+++ bootstrap/app.php
@@ -3,12 +3,20 @@
 use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Configuration\Exceptions;
 use Illuminate\Foundation\Configuration\Middleware;
+use Illuminate\Routing\Middleware\SubstituteBindings;
+use Illuminate\Support\Facades\Route;
 
 return Application::configure(basePath: dirname(__DIR__))
     ->withRouting(
-        web: __DIR__.'/../routes/web.php',
-        commands: __DIR__.'/../routes/console.php',
+        // Using base_path() instead of __DIR__ because it's more uniform.
+        web: base_path('routes/web.php'),
+        api: base_path('routes/api.php'),
+        commands: base_path('routes/console.php'),
         health: '/up',
+        then: function () {
+            Route::middleware(SubstituteBindings::class)
+                ->group(base_path('routes/delivery.php'));
+        },
     )
     ->withMiddleware(function (Middleware $middleware) {
         //
